+++ src/commands/constants.ts
@@ -1,4 +1,6 @@
 export namespace CommandNames {
   export const ImportIdentifierCommandName = 'haskell.commands.importIdentifier';
   export const RestartServerCommandName = 'haskell.commands.restartServer';
+  export const StartServerCommandName = 'haskell.commands.startServer';
+  export const StopServerCommandName = 'haskell.commands.stopServer';
 }