+++ Changelog.md
@@ -2,6 +2,14 @@
 
 - Add tracking of cabal files to work together with the incoming cabal formatter plugin
 
+### 1.5.1
+
+- Add much more logging in the client side, configured with `haskell.trace.client`
+- Fix error handling of `working out project ghc` (See #421)
+  - And dont use a shell to spawn the subprocess in non windows systems
+  - Show the progress as a cancellable notification
+- Add commands `Start Haskell LSP server` and `Stop Haskell LSP server`
+
 ### 1.5.0
 
 - Emit warning about limited support for ghc-9.x on hls executable download